core/exec: add a named-descriptor option for streams
This commit adds a `named-descriptor` option to `StandardInput=`,
`StandardOutput=` and `StandardError=` properties in order to
connect standard streams to externally named descriptors provided
by some socket units.

This option looks for a file descriptor named as the corresponding
property. If multiple name-matches exist, the first matching fd
will be used.