core/exec: accept named sockets for streams
This commit adds a free-form option to `StandardInput=`,
`StandardOutput=` and `StandardError=` properties in order to
connect standard streams to externally named descriptors provided
by some socket units.

This option looks for a file descriptor named as the property value.
If multiple name-matches exist, the first matching fd
will be used.